+++ Assets/Localization/LocalSettings/Tables/PlateCountMethod Shared Data.asset
@@ -35,6 +35,10 @@ MonoBehaviour:
     m_Key: FillTubesPopup
     m_Metadata:
       m_Items: []
+  - m_Id: 29158718955794432
+    m_Key: toolsToCabinet
+    m_Metadata:
+      m_Items: []
   m_Metadata:
     m_Items: []
   m_KeyGenerator:
