+++ Assets/Localization/LocalSettings/Tables/PlateCountMethod_en.asset
@@ -38,6 +38,10 @@ MonoBehaviour:
     m_Localized: Tubes filled!
     m_Metadata:
       m_Items: []
+  - m_Id: 29158718955794432
+    m_Localized: Take the sterilized tools to the laminar flow cabinet.
+    m_Metadata:
+      m_Items: []
   references:
     version: 2
     RefIds: []
