+++ Assets/Localization/LocalSettings/Tables/PlateCountMethod_fi.asset
@@ -38,6 +38,10 @@ MonoBehaviour:
     m_Localized: "Putket t\xE4ytetty!"
     m_Metadata:
       m_Items: []
+  - m_Id: 29158718955794432
+    m_Localized: "Vie steriloidut ty\xF6v\xE4lineet laminaarikaappiin"
+    m_Metadata:
+      m_Items: []
   references:
     version: 2
     RefIds: []
