+++ Assets/Localization/LocalSettings/Tables/PlateCountMethod_sv.asset
@@ -38,6 +38,10 @@ MonoBehaviour:
     m_Localized: "R\xF6ren \xE4r fyllda!"
     m_Metadata:
       m_Items: []
+  - m_Id: 29158718955794432
+    m_Localized: "Ta de sterila verktygen till laminarfl\xF6deskabinetet."
+    m_Metadata:
+      m_Items: []
   references:
     version: 2
     RefIds: []
