+++ Assets/Packages/SimpleHands/Materials/White-Hand.mat
@@ -186,7 +186,7 @@ Material:
     - _GlossMapScale: 0
     - _Glossiness: 0
     - _GlossyReflections: 0
-    - _GlovesOn: 1
+    - _GlovesOn: 0
     - _LightingEnabled: 0
     - _Metallic: 0
     - _Mode: 0
@@ -204,7 +204,7 @@ Material:
     - _SoftParticlesNearFadeDistance: 0
     - _SpecularHighlights: 1
     - _SrcBlend: 1
-    - _StepEdge: 0.6
+    - _StepEdge: 0.05
     - _Surface: 0
     - _WorkflowMode: 1
     - _ZWrite: 1
