+++ Assets/Scripts/PlateCountMethod/CabinetBasePCM_Old.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f185d98014a4acb4999abb0ee38f0fd6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
