+++ Assets/Task Lists/PlateCountMethodTasks.asset
@@ -13,6 +13,13 @@ MonoBehaviour:
   m_Name: PlateCountMethodTasks
   m_EditorClassIdentifier: 
   tasks:
+  - <key>k__BackingField: toolsToCabinet
+    <taskText>k__BackingField: "Vie steriloidut ty\xF6v\xE4lineet laminaarikaappiin"
+    <hint>k__BackingField: "Just kuten siin\xE4 sanotaan, laita ty\xF6kalut kaappiin."
+    <points>k__BackingField: 1
+    <timed>k__BackingField: 0
+    <timeToCompleteTask>k__BackingField: 0
+    <failWhenOutOfTime>k__BackingField: 0
   - <key>k__BackingField: WashHands
     <taskText>k__BackingField: "Pese k\xE4tesi!"
     <hint>k__BackingField: "Suihkauta k\xE4si\xE4si suihkepullolla."
